requirements.txt
discord.py
aiohttp
Flask
python-dotenv
supabase